Refactor makeMoneroTools into it's own file
